refactor(federation): reuse sha256Hex and hoist test hub helpers
- peer-auth.ts: use the existing sha256Hex helper from signing.ts
  instead of inlining crypto.createHash; tighten the body-hash
  comment to explain the "-" sentinel rather than restate the code.
- Hoist tmpPath / startHub / admit out of gossip.test.ts and
  peer-announce.test.ts into hub/test/helpers/hub-setup.ts.
