+++ hub/test/peer-announce.test.ts
@@ -7,70 +7,10 @@
  */
 
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
-import os from "node:os";
-import path from "node:path";
 import fs from "node:fs";
-import type { AddressInfo } from "node:net";
-import type { FastifyInstance } from "fastify";
-import { buildApp } from "../src/server.js";
-import { seedAdminUser } from "./helpers/admin-user.js";
+import { admit, startHub, type Hub } from "./helpers/hub-setup.js";
 import { ingestGossipEntry } from "../src/federation/gossip.js";
 import type { GossipPeerEntry } from "../src/federation/gossip.js";
-import type { Config } from "../src/config.js";
-
-function tmpPath(suffix = "") {
-  return path.join(
-    os.tmpdir(),
-    `poutine-announce-${Date.now()}-${Math.random().toString(36).slice(2)}-${suffix}`,
-  );
-}
-
-interface Hub {
-  app: FastifyInstance;
-  port: number;
-  url: string;
-  keyPath: string;
-  token: string;
-}
-
-async function startHub(id: string): Promise<Hub> {
-  const keyPath = tmpPath(`${id}-key.pem`);
-  const config: Partial<Config> = {
-    databasePath: ":memory:",
-    jwtSecret: `test-${id}`,
-    poutinePrivateKeyPath: keyPath,
-    poutineInstanceId: id,
-    poutineOwnerUsername: `admin-${id}`,
-    navidromeUrl: "http://127.0.0.1:1",
-    navidromeUsername: "x",
-    navidromePassword: "x",
-  };
-  const app = await buildApp(config);
-  await app.ready();
-  await app.listen({ port: 0, host: "127.0.0.1" });
-  const port = (app.server.address() as AddressInfo).port;
-  const url = `http://127.0.0.1:${port}`;
-  const { token } = await seedAdminUser(app, `admin-${id}`);
-  return { app, port, url, keyPath, token };
-}
-
-async function admit(inviter: Hub, invitee: Hub) {
-  const issue = await inviter.app.inject({
-    method: "POST",
-    url: "/admin/peers/invite",
-    headers: { authorization: `Bearer ${inviter.token}` },
-    payload: { ourUrl: inviter.url, inviteeUrl: invitee.url, expiresInSec: 600 },
-  });
-  expect(issue.statusCode).toBe(200);
-  const { invitation } = issue.json() as { invitation: string };
-  const accept = await invitee.app.inject({
-    method: "POST",
-    url: "/admin/peers/accept",
-    headers: { authorization: `Bearer ${invitee.token}` },
-    payload: { invitation, ourUrl: invitee.url },
-  });
-  expect(accept.statusCode).toBe(200);
-}
 
 // Wait until pred() is truthy or `timeoutMs` elapses.
 async function waitFor(pred: () => boolean, timeoutMs = 2000, stepMs = 25) {
